chore: move screen and component and rename sendtip to follow convention
